+++ src/components/ResultSearchCard/SearchInfo/SearchInfo.tsx
@@ -16,7 +16,7 @@ import {
     StyledInfoIcon,
     StyledRemoveIcon,
 } from '../../ListCard/styles'
-import { KeyWords } from '../styles'
+import { KeyWords } from '../../SearchResultCard/styles'
 import {
     DescriptionParagraph,
     FirstInfoBox,
