+++ src/components/ResultSearchCard/SearchInfo/styles.ts
