+++ src/pages/addPart/recentlyAdded/RecentlyAdded.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from 'react-router-dom'
 import { Button } from '../../../components/Button/Button.tsx'
-import SearchResultCardCompact from '../../../components/SearchResultCard/SearchInfoCompact.tsx'
-import SearchResultCard from '../../../components/SearchResultCard/SearchResultCard.tsx'
+import SearchResultCard from '../../../components/ResultSearchCard/ResultSearchCard.tsx'
+import SearchResultCardCompact from '../../../components/ResultSearchCard/SearchInfoCompact.tsx'
 import { useWindowDimensions } from '../../../hooks/index.ts'
 import { Item } from '../../../services/apiTypes.ts'
 import { useGetItemsByUser } from '../../../services/hooks/items/useGetItemByUser.tsx'
