+++ src/pages/admin/category/Categories.tsx
@@ -7,7 +7,7 @@ import AdminSearchCard, {
 import { Button } from '../../../components/Button/Button'
 import SearchBar from '../../../components/SearchBar/SearchBar'
 import { Category } from '../../../services/apiTypes'
-import { useGetCategories } from '../../../services/hooks/category/useGetCategories'
+import { useGetCategories } from '../../../services/hooks/categoryFix/useGetCategories'
 import {
     AdminContainer,
     ButtonContainer,
