+++ src/pages/listDetails/ListDetails.tsx
@@ -2,9 +2,9 @@ import { useContext, useEffect, useState } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useDebounce } from 'usehooks-ts'
 import { Button } from '../../components/Button/Button.tsx'
+import SearchResultCard from '../../components/ResultSearchCard/ResultSearchCard.tsx'
+import SearchResultCardCompact from '../../components/ResultSearchCard/SearchInfoCompact.tsx'
 import SearchBar from '../../components/SearchBar/SearchBar.tsx'
-import SearchResultCardCompact from '../../components/SearchResultCard/SearchInfoCompact.tsx'
-import SearchResultCard from '../../components/SearchResultCard/SearchResultCard.tsx'
 import UmAppContext from '../../contexts/UmAppContext.tsx'
 import { useSnackBar, useWindowDimensions } from '../../hooks'
 import { Item, UpdateList } from '../../services/apiTypes.ts'
