+++ src/pages/listDetails/Phone/List/PhoneList.tsx
@@ -3,12 +3,12 @@ import { useNavigate, useParams } from 'react-router-dom'
 import { useDebounce } from 'usehooks-ts'
 import { Button } from '../../../../components/Button/Button'
 import { GlobalSpinner } from '../../../../components/GlobalSpinner/GlobalSpinner'
-import SearchResultCardCompact from '../../../../components/SearchResultCard/SearchInfoCompact'
+import SearchResultCardCompact from '../../../../components/ResultSearchCard/SearchInfoCompact'
 import UmAppContext from '../../../../contexts/UmAppContext'
 import { useSnackBar } from '../../../../hooks'
 import { Item, List, UpdateList } from '../../../../services/apiTypes'
-import { useGetListById } from '../../../../services/hooks/list/useGetListById'
 import { useGetItemsNotInListInfinite } from '../../../../services/hooks/items/useGetItemsNotInListInfinite'
+import { useGetListById } from '../../../../services/hooks/list/useGetListById'
 import { useUpdateList } from '../../../../services/hooks/list/useUpdateList'
 import { COLORS } from '../../../../style/GlobalStyles'
 import { ListHeader } from '../../ListHeader'
