+++ src/pages/listDetails/phone/AddMore.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { useDebounce } from 'usehooks-ts'
 import { GlobalSpinner } from '../../../components/GlobalSpinner/GlobalSpinner'
+import SearchResultCardCompact from '../../../components/ResultSearchCard/SearchInfoCompact'
 import SearchBar from '../../../components/SearchBar/SearchBar'
-import SearchResultCardCompact from '../../../components/SearchResultCard/SearchInfoCompact'
 import { useSnackBar } from '../../../hooks'
 import { useGetListById } from '../../../services/hooks/list/useGetListById'
