--- src/services/hooks/category/useGetCategories.tsx
@@ -1,9 +1,0 @@
-import { useQuery } from '@tanstack/react-query'
-import apiService from '../../api'
-
-export const useGetCategories = () => {
-    return useQuery({
-        queryKey: ['categories'],
-        queryFn: () => apiService().getCategory(),
-    })
-}
